requirements.txt
python-dateutil==2.2
MySQL-python==1.2.5
PyYAML==3.10
Jinja2==2.6
